Fix: Loading of textures of materials from USDZ
In the case of USDZ, stage textures referenced by materials could be embedded into the stage file which is a zip archive.

If the path to the stage is "/scene.usdz" and "0/image.png" is an asset path referenced from some material then resolved path would be "/scene.usdz[0/image.png]" which definitely not supported by rprContextCreateImageFromFile

So the best solution is to use USD API for loading images because such cases would be handled automatically. Also will be supported all image formats that USD supports, i.e. imaging plugins, such as opeimageio, will automatically expand the list of supported image formats