Remove global framebuffer size
Instead we define the size of each AOV framebuffer separately.
It happens in HdRprRenderBuffer::_Allocate so that's if a user
asks for a particular size he will get it and it will not change.
Previously, it used to be that we could allocate HdRprRenderBuffer
with one size, and then it changed its internal size, unknowingly to the user,
via ResizeFramebuffers that called from HdRprRenderPass::_Execute
based on viewport size