Add sampling controls
This checkin adds some sampling controls.  Namely min samples, max samples, and variance.

We set is_converged on the render buffer and the render pass based on the criteria
 is samples > max samples or are all pixels converged to variance threshold.

Open TODOs:
How do we mark buffers and render pass as not converged and set iterations to 0 when changes happen?
How do we set defaults here?
Expose in Menu?