Update to RPR SDK 1.34.3 (#63)
RPR 1.34.3 added new function - rprFrameBufferFillWithColor - that will allow me to fix a few bugs we have. So as previously we had no ability to fill framebuffer with custom color, we always had to fill it with zero using rprFrameBufferClear. It causes such inconsistencies for pixels that were not hit by ray:

WorldCoordinate AOV (used for ndc depth calculation) - invalid depth value
ObjectID AOV - conform to the first index, but actually should be some sort of invalid index (e.g. -1)

* Remove global framebuffer size

Instead we define the size of each AOV framebuffer separately.
It happens in HdRprRenderBuffer::_Allocate so that's if a user
asks for a particular size he will get it and it will not change.
Previously, it used to be that we could allocate HdRprRenderBuffer
with one size, and then it changed its internal size, unknowingly to the user,
via ResizeFramebuffers that called from HdRprRenderPass::_Execute
based on viewport size

* Update to RPR SDK 1.34.3

* Fix linkage against TBB library

Link only to tbb library instead of linking to all available tbb libraries

* Fixes needed for macOS

macOS should always use METAL.  Also remove -undefined dynamic_lookup as it hid link errors.  Also #ifdef out some calls that are not defined on macOS.

* Fix hdRpr's install name of RPR library

* Update to RPR SDK 1.34.3

* use RPR_MATERIAL_NODE_NORMAL_MAP for normal inputs.