Fix iteration count initialization
After the first hdRpr initialization, HdRprConfig's sampling attributes were marked as clean. But when you reinitialize hdRpr, HdRprConfig's sampling is no longer dirty that's why the number of max samples was never queried again until you set a new value for one of the sampling attributes in HdRprConfig. Thus the number of max samples was always 0. This behavior was mentioned in #95