Fix querying of UsdPreviewSurface textures
In the case when the same texture used for two relationships, the second one was ignored.
For example, we have UsdUvTexture referencing png with transparency. In that case, at the same time this one texture may be used as diffuseColor with `rgb` inputName and as opacity map with `a` inputName. Previously, opacity relationship was ignored.