Fix undefined behavior on abort of the first rprContextRender
Ideally, aborting the first sample should not be the problem. We would like to be able to abort it: to reduce response time, to avoid doing calculations that may be discarded by the following changes.
But in reality, aborting the very first sample may cause crashes or unwanted behavior when we will call rprContextRender next time.
So until RPR core fixes these issues, we are not aborting the first sample.