Make .rpr export node work with any character in the output path
We are using output filename as a part of the path to dummy cube that forces render when this node is applied:
```
stage = node.editableStage()
cube = stage.DefinePrim('/rprExportImplicitPrimitive/' + filename, 'Cube')
```
where `filename, ext = os.path.splitext(os.path.basename(exportPath))`

So when `filename` contained any symbol that was not valid for SdfPath, the export node was getting an error.
To fix it, we preprocess filename with Tf.MakeValidIdentifier to remove all possibly invalid characters from it, so that it can be used as part of SdfPath.