Fix depth AOV
Depth AOV always had wrong values: on the very first render pass it was zeroed, after that depth values were always one frame in the past.
Such behavior is hardly noticeable in the IPR but a quite obvious when with cmd-line rendering.
This was caused by the wrong ordering of rif::Filter::Update calls.
rif::Filter::Update function attaches the filter to the context's command queue.
The remap filter was attached before the NDC filter.