Fix rendering when color AOV is not bound
Since #413, color AOV is not created in advance.
From the other side, RPR refuses to render if color AOV is not created.
It's possible for Hydra to ignore color AOV.
For example, in usdview when depth AOV is selected, only depth AOV is bound.
That's why we must always create color AOV.